♻️ refactor: button type 폴더 분리
