+++ .eslintrc
@@ -30,6 +30,7 @@
       "off",
       { "rootDir": "src", "allowSameFolder": true }
     ],
+    "import/no-cycle": "error",
     "import/no-unused-modules": [
       "off",
       {
